+++ lib/nosql/secure_storage/secure_storage_service.dart
@@ -7,11 +7,15 @@ abstract interface class SecureStorageService {
 
   Future<void> write(String key, String value);
 
+  Future<void> writeBatch(Map<String, String> data);
+
   Future<String?> read(String key);
 
   Future<bool> containsKey(String key);
 
   Future<Map<String, String>> readAll();
+
+  Future<List<String>> getKeys();
 
   Future<void> delete(String key);